+++ src/main/java/nora/movlog/controller/MovieController.java
@@ -2,7 +2,7 @@ package nora.movlog.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nora.movlog.entity.Movie;
+import nora.movlog.domain.Movie;
 import nora.movlog.service.MovieService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
